+++ Sources/Schemas/OrderExpanded.swift
@@ -6,6 +6,7 @@ public struct OrderExpanded: Codable, Hashable, Sendable {
     public let forwarderOrgHandle: String?
     public let shipperOrgCompanyName: String?
     public let offChrtShipperOrgCompanyName: String?
+    public let notificationIntentsAdHoc: [NotificationIntentAdHoc1]?
     public let taskGroupsExpanded: [TaskGroupExpanded]?
     /// Additional properties that are not explicitly defined in the schema
     public let additionalProperties: [String: JSONValue]
@@ -16,6 +17,7 @@ public struct OrderExpanded: Codable, Hashable, Sendable {
         forwarderOrgHandle: String? = nil,
         shipperOrgCompanyName: String? = nil,
         offChrtShipperOrgCompanyName: String? = nil,
+        notificationIntentsAdHoc: [NotificationIntentAdHoc1]? = nil,
         taskGroupsExpanded: [TaskGroupExpanded]? = nil,
         additionalProperties: [String: JSONValue] = .init()
     ) {
@@ -24,6 +26,7 @@ public struct OrderExpanded: Codable, Hashable, Sendable {
         self.forwarderOrgHandle = forwarderOrgHandle
         self.shipperOrgCompanyName = shipperOrgCompanyName
         self.offChrtShipperOrgCompanyName = offChrtShipperOrgCompanyName
+        self.notificationIntentsAdHoc = notificationIntentsAdHoc
         self.taskGroupsExpanded = taskGroupsExpanded
         self.additionalProperties = additionalProperties
     }
@@ -35,6 +38,7 @@ public struct OrderExpanded: Codable, Hashable, Sendable {
         self.forwarderOrgHandle = try container.decodeIfPresent(String.self, forKey: .forwarderOrgHandle)
         self.shipperOrgCompanyName = try container.decodeIfPresent(String.self, forKey: .shipperOrgCompanyName)
         self.offChrtShipperOrgCompanyName = try container.decodeIfPresent(String.self, forKey: .offChrtShipperOrgCompanyName)
+        self.notificationIntentsAdHoc = try container.decodeIfPresent([NotificationIntentAdHoc1].self, forKey: .notificationIntentsAdHoc)
         self.taskGroupsExpanded = try container.decodeIfPresent([TaskGroupExpanded].self, forKey: .taskGroupsExpanded)
         self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
     }
@@ -47,6 +51,7 @@ public struct OrderExpanded: Codable, Hashable, Sendable {
         try container.encodeIfPresent(self.forwarderOrgHandle, forKey: .forwarderOrgHandle)
         try container.encodeIfPresent(self.shipperOrgCompanyName, forKey: .shipperOrgCompanyName)
         try container.encodeIfPresent(self.offChrtShipperOrgCompanyName, forKey: .offChrtShipperOrgCompanyName)
+        try container.encodeIfPresent(self.notificationIntentsAdHoc, forKey: .notificationIntentsAdHoc)
         try container.encodeIfPresent(self.taskGroupsExpanded, forKey: .taskGroupsExpanded)
     }
 
@@ -57,6 +62,7 @@ public struct OrderExpanded: Codable, Hashable, Sendable {
         case forwarderOrgHandle = "forwarder_org_handle"
         case shipperOrgCompanyName = "shipper_org_company_name"
         case offChrtShipperOrgCompanyName = "off_chrt_shipper_org_company_name"
+        case notificationIntentsAdHoc = "notification_intents_ad_hoc"
         case taskGroupsExpanded = "task_groups_expanded"
     }
 }