+++ Sources/Schemas/StartOfWeek.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public enum StartOfWeek: String, Codable, Hashable, CaseIterable, Sendable {
+    case monday
+    case tuesday
+    case wednesday
+    case thursday
+    case friday
+    case saturday
+    case sunday
+}