+++ Sources/Schemas/TimeBucket.swift
@@ -1,0 +1,38 @@
+import Foundation
+
+public struct TimeBucket: Codable, Hashable, Sendable {
+    public let period: Date
+    public let value: Double
+    /// Additional properties that are not explicitly defined in the schema
+    public let additionalProperties: [String: JSONValue]
+
+    public init(
+        period: Date,
+        value: Double,
+        additionalProperties: [String: JSONValue] = .init()
+    ) {
+        self.period = period
+        self.value = value
+        self.additionalProperties = additionalProperties
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        self.period = try container.decode(Date.self, forKey: .period)
+        self.value = try container.decode(Double.self, forKey: .value)
+        self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
+    }
+
+    public func encode(to encoder: Encoder) throws -> Void {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try encoder.encodeAdditionalProperties(self.additionalProperties)
+        try container.encode(self.period, forKey: .period)
+        try container.encode(self.value, forKey: .value)
+    }
+
+    /// Keys for encoding/decoding struct properties.
+    enum CodingKeys: String, CodingKey, CaseIterable {
+        case period
+        case value
+    }
+}