+++ Sources/Schemas/UserPrivateData1.swift
@@ -1,0 +1,68 @@
+import Foundation
+
+public struct UserPrivateData1: Codable, Hashable, Sendable {
+    public let schemaVersion: Int
+    public let createdAtTimestamp: Date
+    public let userId: String
+    public let stripeCustomerId: String?
+    public let primaryEmailAddress: String?
+    public let phoneNumber: String?
+    public let id: String
+    /// Additional properties that are not explicitly defined in the schema
+    public let additionalProperties: [String: JSONValue]
+
+    public init(
+        schemaVersion: Int,
+        createdAtTimestamp: Date,
+        userId: String,
+        stripeCustomerId: String? = nil,
+        primaryEmailAddress: String? = nil,
+        phoneNumber: String? = nil,
+        id: String,
+        additionalProperties: [String: JSONValue] = .init()
+    ) {
+        self.schemaVersion = schemaVersion
+        self.createdAtTimestamp = createdAtTimestamp
+        self.userId = userId
+        self.stripeCustomerId = stripeCustomerId
+        self.primaryEmailAddress = primaryEmailAddress
+        self.phoneNumber = phoneNumber
+        self.id = id
+        self.additionalProperties = additionalProperties
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        self.schemaVersion = try container.decode(Int.self, forKey: .schemaVersion)
+        self.createdAtTimestamp = try container.decode(Date.self, forKey: .createdAtTimestamp)
+        self.userId = try container.decode(String.self, forKey: .userId)
+        self.stripeCustomerId = try container.decodeIfPresent(String.self, forKey: .stripeCustomerId)
+        self.primaryEmailAddress = try container.decodeIfPresent(String.self, forKey: .primaryEmailAddress)
+        self.phoneNumber = try container.decodeIfPresent(String.self, forKey: .phoneNumber)
+        self.id = try container.decode(String.self, forKey: .id)
+        self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
+    }
+
+    public func encode(to encoder: Encoder) throws -> Void {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try encoder.encodeAdditionalProperties(self.additionalProperties)
+        try container.encode(self.schemaVersion, forKey: .schemaVersion)
+        try container.encode(self.createdAtTimestamp, forKey: .createdAtTimestamp)
+        try container.encode(self.userId, forKey: .userId)
+        try container.encodeIfPresent(self.stripeCustomerId, forKey: .stripeCustomerId)
+        try container.encodeIfPresent(self.primaryEmailAddress, forKey: .primaryEmailAddress)
+        try container.encodeIfPresent(self.phoneNumber, forKey: .phoneNumber)
+        try container.encode(self.id, forKey: .id)
+    }
+
+    /// Keys for encoding/decoding struct properties.
+    enum CodingKeys: String, CodingKey, CaseIterable {
+        case schemaVersion = "schema_version"
+        case createdAtTimestamp = "created_at_timestamp"
+        case userId = "user_id"
+        case stripeCustomerId = "stripe_customer_id"
+        case primaryEmailAddress = "primary_email_address"
+        case phoneNumber = "phone_number"
+        case id = "_id"
+    }
+}